Tests: Move utils emails tests to their own module
